Add teams folder and shared tier select. Update tier select use and styling in invite step
